# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API document of Watchlist Information System",
        "title": "Watchlist Information System API",
        "termsOfService": "COJ-App Team",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "10.1.2.124:9005",
    "basePath": "/api/v1",
    "paths": {
...
        "/setting": {
            "get": {
                "description": "เรียกดูรายการตั้งค่าหนังสือของศาล",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ตั้งค่าหนังสือ"
                ],
                "summary": "เรียกดูรายการตั้งค่าหนังสือ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "รหัสศาล",
                        "name": "court_code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DataSetting"
                        }
                    }
                }
            },
            "post": {
                "description": "เพิ่มข้อมูลการตั้งค่าหนังสือ \u003cbr\u003e*** Required \u003cbr\u003e\"court_code\", \"setting_name\", \"created_uid\", \"created_uname\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ตั้งค่าหนังสือ"
                ],
                "summary": "เพิ่มข้อมูลการตั้งค่าหนังสือ",
                "parameters": [
                    {
                        "description": "Create Setting",
                        "name": "setting",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                }
            }
        },
        "/setting/{id}": {
            "get": {
                "description": "ดูข้อมูลการตั้งค่าหนังสือของศาล",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ตั้งค่าหนังสือ"
                ],
                "summary": "ดูข้อมูลการตั้งค่าหนังสือ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                }
            },
            "put": {
                "description": "แก้ไขข้อมูลการตั่งค่าหนังสือ \u003cbr\u003e*** Required \u003cbr\u003e\"updated_uid\", \"updated_uname\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ตั้งค่าหนังสือ"
                ],
                "summary": "แก้ไขข้อมูลการตั้งค่าหนังสือ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Setting",
                        "name": "setting",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                }
            },
            "delete": {
                "description": "ลบข้อมูลการตั่งค่าหนังสือ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ตั้งค่าหนังสือ"
                ],
                "summary": "ลบข้อมูลการตั้งค่าหนังสือ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DataSetting": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Setting"
                    }
                },
                "status": {
                    "description": "รหัสสถานะการเรียก service",
                    "type": "integer",
                    "example": 200
                },
                "total": {
                    "description": "จำนวน Rows ทั้งหมด",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.DefaultMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message example"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.Setting": {
            "type": "object",
            "properties": {
                "court_addr": {
                    "description": "ที่อยู่ของศาล",
                    "type": "string",
                    "example": "อาคารศาลอาญา ถนนรัชดาภิเษก แขวงจอมพล เขตจตุจักร กรุงเทพฯ 10900"
                },
                "court_code": {
                    "description": "รหัสศาล",
                    "type": "string",
                    "example": "0000011"
                },
                "court_name": {
                    "description": "ชื่อศาล",
                    "type": "string",
                    "example": "สำนักเทคโนโลยีสารสนเทศ"
                },
                "created_at": {
                    "description": "วันที่สร้างข้อมูล",
                    "type": "string",
                    "example": "2019-10-15T14:01:58+07:00"
                },
                "created_uid": {
                    "description": "ID ผู้สร้างข้อมูล",
                    "type": "integer",
                    "example": 1
                },
                "created_uname": {
                    "description": "ชื่อผู้สร้างข้อมูล",
                    "type": "string",
                    "example": "user01"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "letter_court_num": {
                    "description": "เลขที่หนังสือ ศย",
                    "type": "string",
                    "example": "ศย102/62"
                },
                "letter_email": {
                    "description": "อีเมล์ในหนังสือ",
                    "type": "string",
                    "example": "mail@coj.go.th"
                },
                "letter_fax": {
                    "description": "แฟกซ์ในหนังสือ",
                    "type": "string",
                    "example": "0266666"
                },
                "letter_group_name": {
                    "description": "ชื่อกลุ่มงานในหนังสือ",
                    "type": "string",
                    "example": "ระบบงาน"
                },
                "letter_office_name": {
                    "description": "ชื่อหน่วยงานในหนังสือ",
                    "type": "string",
                    "example": "สำนักเทคโนโลยีสารสนเทศ"
                },
                "letter_signed_name": {
                    "description": "ชื่อผู้เซ็นในหนังสือ",
                    "type": "string",
                    "example": "นายลายเซ็น หนังสือ"
                },
                "letter_signed_position": {
                    "description": "ตำแหน่งผู้เซ็นในหนังสือ",
                    "type": "string",
                    "example": "นักวิจัดการงานทั่วไปชำนาญการ"
                },
                "letter_tel": {
                    "description": "เบอร์โทรในหนังสือ",
                    "type": "string",
                    "example": "0255555"
                },
                "setting_name": {
                    "description": "ชื่อการตั้งค่าหนังสือ",
                    "type": "string",
                    "example": "ตั้งค่า 1"
                },
                "updated_at": {
                    "description": "วันที่แก้ไขข้อมูลล่าสุด",
                    "type": "string",
                    "example": "2019-10-15T14:01:58+07:00"
                },
                "updated_uid": {
                    "description": "ID ผู้แก้ไขข้อมูลล่าสุด",
                    "type": "integer",
                    "example": 1
                },
                "updated_uname": {
                    "description": "ชื่อผู้แก้ไขข้อมูลล่าสุด",
                    "type": "string",
                    "example": "user01"
                }
            }
        },
...
    }
}